extract routes, io and server logic from server.js
